+++ src/tap/algorithms/discrete_filter.hpp
@@ -23,10 +23,26 @@
 #include <array>
 #include <cstdint>
 
-namespace tap
+namespace tap::algorithms::filter
 {
-namespace algorithms
+/**
+ * @struct Coefficients
+ * @brief Represents the coefficients used in a discrete filter.
+ *
+ * This structure holds two sets of coefficients:
+ * - `naturalResponseCoefficients`: Coefficients related to the natural response of the system.
+ * - `forcedResponseCoefficients`: Coefficients related to the forced response of the system.
+ *
+ * @tparam T The data type of the coefficients (e.g., float, double).
+ * @tparam SIZE The size of the coefficient arrays.
+ */
+template <uint8_t SIZE, typename T = float>
+struct Coefficients
 {
+    std::array<T, SIZE> naturalResponseCoefficients;
+    std::array<T, SIZE> forcedResponseCoefficients;
+};
+
 /**
  * @brief DiscreteFilter class implements a discrete-time filter using the finite difference
  * equation.
@@ -48,10 +64,17 @@ public:
      * state to zero.
      */
     DiscreteFilter(
-        std::array<T, SIZE> &naturalResponseCoefficients,
-        std::array<T, SIZE> &forcedResponseCoefficients)
+        std::array<T, SIZE> naturalResponseCoefficients,
+        std::array<T, SIZE> forcedResponseCoefficients)
         : naturalResponseCoefficients(naturalResponseCoefficients),
           forcedResponseCoefficients(forcedResponseCoefficients)
+    {
+        reset();
+    }
+
+    DiscreteFilter(const Coefficients<SIZE, T> coefficients)
+        : naturalResponseCoefficients(coefficients.naturalResponseCoefficients),
+          forcedResponseCoefficients(coefficients.forcedResponseCoefficients)
     {
         reset();
     }
@@ -117,6 +140,20 @@ public:
         return 0.0f;
     }
 
+    void setCoefficients(Coefficients<SIZE, T> coe)
+    {
+        this->naturalResponseCoefficients = coe.naturalResponseCoefficients;
+        this->forcedResponseCoefficients = coe.forcedResponseCoefficients;
+    }
+
+    void setCoefficients(
+        std::array<T, SIZE> naturalResponseCoefficients,
+        std::array<T, SIZE> forcedResponseCoefficients)
+    {
+        this->naturalResponseCoefficients = naturalResponseCoefficients;
+        this->forcedResponseCoefficients = forcedResponseCoefficients;
+    }
+
 private:
     std::array<T, SIZE> naturalResponseCoefficients;
     std::array<T, SIZE> forcedResponseCoefficients;
@@ -124,8 +161,6 @@ private:
     std::array<T, SIZE> forcedResponse;
 };
 
-}  // namespace algorithms
-
-}  // namespace tap
+}  // namespace tap::algorithms::filter
 
 #endif  // TAPROOT_DISCRETE_FILTER_HPP_