+++ test/tap/mock/odometry_2d_interface_mock.hpp
@@ -38,6 +38,7 @@ public:
         overrideOdometryPosition,
         (const float positionX, const float positionY),
         (override));
+    MOCK_METHOD(void, reset, (), (override));
 };
 }  // namespace tap::mock